LGVISIUM-102: merge AToBInterval and AToBDepthColumnEntry
