+++ src/stratigraphy/depthcolumnentry/__init__.py
@@ -1,0 +1,6 @@
+"""Modules for depth column entries, values that indicate the measured depth of an interface between layers."""
+
+from .depthcolumnentry import DepthColumnEntry
+from .depthcolumnentry_extractor import DepthColumnEntryExtractor
+
+__all__ = ["DepthColumnEntry", "DepthColumnEntryExtractor"]
