+++ src/stratigraphy/layer/layer.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 
 import fitz
 from stratigraphy.data_extractor.data_extractor import ExtractedFeature, FeatureOnPage
-from stratigraphy.depthcolumn.depthcolumnentry import DepthColumnEntry
+from stratigraphy.depthcolumnentry import DepthColumnEntry
 from stratigraphy.text.textblock import MaterialDescription, TextBlock
 from stratigraphy.util.interval import AAboveBInterval, Interval
 from stratigraphy.util.util import parse_text
