+++ src/stratigraphy/sidebar/a_above_b_sidebar_extractor.py
@@ -2,7 +2,7 @@
 
 import fitz
 
-from stratigraphy.depthcolumn.depthcolumnentry import DepthColumnEntry
+from stratigraphy.depthcolumnentry import DepthColumnEntryExtractor
 from stratigraphy.lines.line import TextWord
 from stratigraphy.sidebar.a_above_b_sidebar import AAboveBSidebar
 from stratigraphy.sidebar.a_above_b_sidebar_validator import AAboveBSidebarValidator
@@ -27,7 +27,7 @@ class AAboveBSidebarExtractor:
         """
         entries = [
             entry
-            for entry in DepthColumnEntry.find_in_words(all_words, include_splits=False)
+            for entry in DepthColumnEntryExtractor.find_in_words(all_words, include_splits=False)
             if entry.rect not in used_entry_rects
         ]
