+++ src/stratigraphy/util/interval.py
@@ -6,10 +6,7 @@ import abc
 
 import fitz
 
-from stratigraphy.depthcolumn.depthcolumnentry import (
-    AToBDepthColumnEntry,
-    DepthColumnEntry,
-)
+from stratigraphy.depthcolumnentry import DepthColumnEntry
 from stratigraphy.lines.line import TextLine
 from stratigraphy.text.textblock import TextBlock
 
@@ -143,9 +140,13 @@ class AAboveBInterval(Interval):
 class AToBInterval(Interval):
     """Class for intervals that are defined in a single line like "1.00 - 2.30m"."""
 
-    def __init__(self, layer_depth_column_entry: AToBDepthColumnEntry):
-        self.entry = layer_depth_column_entry
-        super().__init__(layer_depth_column_entry.start, layer_depth_column_entry.end)
+    def __init__(self, start: DepthColumnEntry, end: DepthColumnEntry):
+        super().__init__(start, end)
+
+    @property
+    def rect(self) -> fitz.Rect:
+        """Get the rectangle surrounding the interval."""
+        return fitz.Rect(self.start.rect).include_rect(self.end.rect)
 
     @property
     def line_anchor(self) -> fitz.Point | None:
@@ -177,38 +178,3 @@ class AToBInterval(Interval):
             return [TextBlock(matched_lines)]
         else:
             return []
-
-    @classmethod
-    def get_depth_interval_from_lines(cls, lines: list[TextLine]) -> AToBInterval | None:
-        """Extract depth interval from text lines.
-
-        For borehole profiles in the Deriaz layout, the depth interval is usually found in the text of the material
-        description. Often, these text descriptions contain a further separation into multiple sub layers.
-        These sub layers have their own depth intervals. This function extracts the overall depth interval,
-        spanning across all mentioned sub layers.
-
-        Args:
-            lines (list[TextLine]): The lines to extract the depth interval from.
-
-        Returns:
-            AToBInterval | None: The depth interval (if any) or None (if no depth interval was found).
-        """
-        depth_entries = []
-        for line in lines:
-            try:
-                layer_depth_entry = AToBDepthColumnEntry.from_text(line.text, line.rect, require_start_of_string=False)
-                # require_start_of_string = False because the depth interval may not always start at the beginning
-                # of the line e.g. "Remblais Heterogene: 0.00 - 0.5m"
-                if layer_depth_entry:
-                    depth_entries.append(layer_depth_entry)
-            except ValueError:
-                pass
-
-        if depth_entries:
-            # Merge the sub layers into one depth interval.
-            start = min([entry.start for entry in depth_entries], key=lambda start_entry: start_entry.value)
-            end = max([entry.end for entry in depth_entries], key=lambda end_entry: end_entry.value)
-
-            return AToBInterval(AToBDepthColumnEntry(start, end))
-        else:
-            return None
