+++ tests/test_depthcolumn.py
@@ -1,7 +1,7 @@
 """Test suite for the find_depth_columns module."""
 
 import fitz
-from stratigraphy.depthcolumn.depthcolumnentry import DepthColumnEntry
+from stratigraphy.depthcolumnentry.depthcolumnentry import DepthColumnEntry
 from stratigraphy.sidebar import AAboveBSidebar
